requirements.txt
# yt-agentic-rag requirements
# Directory: yt-agentic-rag/requirements.txt

# =============================================================================
# Core Framework
# =============================================================================
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# =============================================================================
# Database (Supabase with pgvector)
# =============================================================================
supabase>=2.0.0

# =============================================================================
# AI Providers
# =============================================================================
openai>=1.0.0
anthropic>=0.18.0

# =============================================================================
# Google API (for Agentic Tools - Calendar, Gmail)
# =============================================================================
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1

# =============================================================================
# Utilities
# =============================================================================
rich>=13.0.0
